requirements.txt
beautifulsoup4==4.13.3
fastapi==0.115.11
groq==0.18.0
protobuf==6.30.0
pydantic==2.10.6
python-dotenv==1.0.1
selenium==4.29.0
uvicorn==0.34.0